# transport/api/openapi.yaml
openapi: 3.0.3
info:
  title: Newsletter API
  version: 1.0.1
  description: API for managing newsletters, subscriptions, and posts.
servers:
  - url: http://localhost:8080/api/v1
    description: API Server for most endpoints
paths:
...
  /editor/refresh:
    post:
      security:
        - BearerAuth: [ ]
      summary: Refreshes the session
      operationId: refresh
      requestBody:
        description: Refresh token
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupabaseRefreshPayload'
      responses:
        '200':
          description: Session refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '400':
          description: Error reading request body or validating request body
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Response'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Response'
  /editor/showJWT:
    get:
      summary: Show JWT token page from fragment in browser
      operationId: JWTPage
      responses:
        '200':
          description: JWT page retrieved successfully
        '500':
          description: Internal server error - missing HTML file
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Response'
...
components:
  schemas:
...
    SupabaseRefreshPayload:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
...
    Response:
      type: object
      properties:
        message:
          type: string
          description: The response message